Move `Func` to its own file
